% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check frozen dependencies for changes}
\usage{
check_dependencies(freeze_path = dirname(Sys.getenv("FREEZR_DESTINATION")))
}
\description{
Check frozen dependencies for changes
}
