% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr.R
\name{create_empty_project}
\alias{create_empty_project}
\title{Create a template project suitable for use with freezr.}
\usage{
create_empty_project(directory)
}
\description{
Create a template project suitable for use with freezr.
}
