% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr_functions.R
\name{inventory}
\alias{inventory}
\title{Keep track of important items from previous analyses.}
\usage{
inventory(inv_location = NULL, tag = NULL, filename = NULL, extra = "",
  parent_tag = "", delete = FALSE, return_all = FALSE)
}
\arguments{
\item{inv_location}{Path to the inventory you want to create, access, or modify. If possible, this arg
defaults to the parent of the last destination given to `freeze`.}

\item{tag}{identifier for an inventory record that you want to add, access, or modify.}

\item{filename}{relative path from \code{inv_location} for a file that you want to add to the inventory.
Used only when adding or modifying records.}

\item{extra}{Any character string without tabs. Meant for metadata to be associated with the given \code{tag}
and \code{filename}.
This may just be notes, or you could include an underscore- and pipe-delimited list of key-value pairs;
the sky's the limit. Used only when adding or modifying records.}

\item{parent_tag}{identifier for a file that this analysis depends on.}

\item{delete}{Overwrite or delete any existing entry with the given tag. Default is \code{FALSE}.}

\item{return_all}{When retrieving data, return the whole inventory record instead of just the \code{filename}.
Default is \code{FALSE}.}
}
\description{
Keep track of important items from previous analyses.
}
\details{
\code{inventory} is a general tool to organize data as it passes through multiple stages of analysis.
 The central data structure is a table with the filename \code{.inventory.txt}. It has five
 columns: \code{tag}, \code{parent_tag}, \code{date_modified}, \code{filename}, and \code{extra}.

\code{inventory} looks for this table at the \code{inv_location} you specify and creates a new one if needed.
\code{inventory} can do three things:
 \itemize{
 \item{ \strong{Show the whole table} } { If no tag is given, \code{inventory} will return the table. }
 \item{ \strong{Retrieve (delete) a row} } { If there's a \code{tag} given, but no \code{filename}, any
 record matching that tag will be returned (or, if \code{delete==TRUE}, will be deleted). }
 \item{ \strong{Add (modify) a row} } { If both tag and filename are given, an entry with that tag
 will be added. If that tag is present already and \code{delete==TRUE}, the record will be
 altered. If the tag is present but \code{delete==FALSE}, the tag will be altered via \code{make.unique}
 and a new record will be created.}
}
}
