% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_add}
\alias{inventory_add}
\title{Keep track of important items from previous analyses.}
\usage{
inventory_add(tag = NULL, inv_location = NULL, filename = NULL,
  extra = "", parent_tag = "", force = FALSE)
}
\arguments{
\item{tag}{identifier for an inventory record that you want to add.}

\item{inv_location}{Path to the inventory you want to create or modify. If possible, this arg
defaults to the parent of the last destination given to `freeze`.}

\item{filename}{relative path from \code{inv_location} for a file that you want to add to the inventory.}

\item{extra}{Any character string without tabs. Meant for metadata to be associated with the given \code{tag}
and \code{filename}.
This may just be notes, or you could include an underscore- and pipe-delimited list of key-value pairs;
the sky's the limit.}

\item{parent_tag}{identifier for a file that this analysis depends on.}

\item{force}{Overwrite or delete any existing entry with the given tag. Default is \code{FALSE}.}
}
\description{
Keep track of important items from previous analyses.
}
\details{
\code{inventory_*} functions help track data as it passes through multiple stages of analysis.
 The central data structure is a table with the filename \code{.inventory.txt}. It has five
 columns: \code{tag}, \code{parent_tag}, \code{date_modified}, \code{filename}, and \code{extra}.

\code{inventory_add} looks for this table at the \code{inv_location} you specify and creates a new one if needed.
\code{inventory_add} will add a row with the given tag. If that tag is present already and
 \code{force==TRUE}, the record will be altered. If the tag is present but \code{force==FALSE},
 the tag will be altered via \code{make.unique} and a new record will be created.
}
