% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_check}
\alias{inventory_check}
\title{Make sure everything in the inventory is actually present}
\usage{
inventory_check(inv_location = NULL)
}
\arguments{
\item{inv_location}{Where to look for the inventory.

This function issues warnings if the inventory or any file it points to is absent. 
It returns a dataframe with details, unless the inventory is absent, in which case
it returns NULL.}
}
\description{
Make sure everything in the inventory is actually present
}
