% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_exists}
\alias{inventory_exists}
\title{Locate your inventory.}
\usage{
inventory_exists(inv_location = NULL)
}
\arguments{
\item{inv_location}{Path to the inventory you think might exist, or to its parent folder, or to
any folder above that (as long as none of them contain other inventories). When possible, this arg
defaults to the parent of the last destination given to `freeze`.}
}
\description{
Locate your inventory.
}
\details{
\code{inventory_*} functions help organize data as it passes through multiple stages of analysis.
 The central data structure is a table with the filename \code{.inventory.txt}. It has five
 columns: \code{tag}, \code{parent_tag}, \code{date_modified}, \code{filename}, and \code{extra}.

\code{inventory_exists} assesses whether an inventory is accessible from the specified location.
}
