% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_make}
\alias{inventory_make}
\title{Make a new inventory.}
\usage{
inventory_make(inv_location)
}
\arguments{
\item{inv_location}{Path to the inventory you want to create, access, or modify. If possible, this arg
defaults to the parent of the last destination given to `freeze`.}
}
\description{
Make a new inventory.
}
\details{
\code{inventory_*} functions help organize data as it passes through multiple stages of analysis.
 The central data structure is a table with the filename \code{.inventory.txt}. It has five
 columns: \code{tag}, \code{parent_tag}, \code{date_modified}, \code{filename}, and \code{extra}.

\code{inventory_make} will make an inventory at the specified location unless there's one already.
}
