% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_rm}
\alias{inventory_rm}
\title{Remove no-longer important items from the inventory.}
\usage{
inventory_rm(tag = NULL, inv_location = NULL)
}
\arguments{
\item{tag}{identifier for an inventory record that you want to add.}

\item{inv_location}{Path to the inventory you want to create or modify. If possible, this arg
defaults to the parent of the last destination given to `freeze`.}
}
\description{
Remove no-longer important items from the inventory.
}
\details{
\code{inventory_*} functions help track data as it passes through multiple stages of analysis.
 The central data structure is a table with the filename \code{.inventory.txt}. It has five
 columns: \code{tag}, \code{parent_tag}, \code{date_modified}, \code{filename}, and \code{extra}.

\code{inventory_rm} will remove the row with the given tag.
}
