% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr_functions.R, R/inventory.R
\name{inventory_show}
\alias{inventory_show}
\alias{inventory_show}
\title{Return an entire inventory.}
\usage{
inventory_show(inv_location = NULL, make_new = FALSE)

inventory_show(inv_location = NULL, make_new = FALSE)
}
\arguments{
\item{inv_location}{Path to the inventory you want to create, access, or modify. If possible, this arg
defaults to the parent of the last destination given to `freeze`.}

\item{make_new}{If \code{make_new==TRUE} and if \code{inv_location} is empty, a new inventory will be created.}

\item{make_new}{If \code{make_new==TRUE} and if \code{inv_location} is empty, a new inventory will be created.}

\item{inv_location}{Path to the inventory you want to create, access, or modify. If possible, this arg
defaults to the parent of the last destination given to `freeze`.}
}
\description{
Return an entire inventory.

Return an entire inventory.
}
\details{
\code{inventory_*} functions help organize data as it passes through multiple stages of analysis.
 The central data structure is a table with the filename \code{.inventory.txt}. It has five
 columns: \code{tag}, \code{parent_tag}, \code{date_modified}, \code{filename}, and \code{extra}.

\code{inventory_show} will return the whole table.

\code{inventory_*} functions help organize data as it passes through multiple stages of analysis.
 The central data structure is a table with the filename \code{.inventory.txt}. It has five
 columns: \code{tag}, \code{parent_tag}, \code{date_modified}, \code{filename}, and \code{extra}.

\code{inventory_show} will return the whole table.
}
