% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_transfer}
\alias{inventory_transfer}
\title{Copy all files listed in the inventory to the specified target location.}
\usage{
inventory_transfer(inv_location = NULL, target_location, overwrite = F)
}
\arguments{
\item{inv_location}{Where to look for the inventory.}

\item{target_location}{The stuff gets put in <target_location>/inventory. 
To avoid copying ridiculous long file paths, files are renamed as "tag.ext" so that 
the new filename is the tag but the old extension (anything after the last period) is preserved.}

\item{overwrite}{Passed to file.copy and also checked before (over)writing the new .inventory file.}
}
\description{
Copy all files listed in the inventory to the specified target location.
}
