% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr_functions.R
\name{pad_list}
\alias{pad_list}
\title{Turn a ragged list of atomic vectors to a rectangle by adding filler.}
\usage{
pad_list(x, filler = "")
}
\description{
Turn a ragged list of atomic vectors to a rectangle by adding filler.
}

