% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr.R, R/freezr_functions.R
\name{thaw}
\alias{thaw}
\alias{thaw}
\title{Run an analysis previously frozen by \code{freezr::freeze}. NOT GUARANTEED TO WORK but it does its best.}
\usage{
thaw(freeze_path, alter_dependencies = F)

thaw(freeze_path, alter_dependencies = F)
}
\arguments{
\item{freeze_path}{A directory created by \code{freezr::freeze} containing these subfolders:
- logs
- code
- output (optional)
- dependencies (optional)
- user (optional)}

\item{use_old_dependencies}{Defaults to \code{FALSE}, so if your dependencies have changed, then
your analysis may not come out of cryo storage working exactly the same way.
If \code{alter_dependencies==TRUE}, then before reviving your analysis, \code{thaw} will GO INTO
THE FILES AND FOLDERS LISTED AS DEPENDENCIES, MOVE THEM OUT OF THE WAY, AND PUT THE FROZEN VERSIONS IN THEIR PLACES.
It cleans up after itself, though, moving the newer dependencies back once it has finished.}

\item{freeze_path}{A directory created by \code{freezr::freeze} containing these subfolders:
- logs
- code
- output (optional)
- dependencies (optional)
- user (optional)}

\item{use_old_dependencies}{Defaults to \code{FALSE}, so if your dependencies have changed, then
your analysis may not come out of cryo storage working exactly the same way.
If \code{alter_dependencies==TRUE}, then before reviving your analysis, \code{thaw} will GO INTO
THE FILES AND FOLDERS LISTED AS DEPENDENCIES, MOVE THEM OUT OF THE WAY, AND PUT THE FROZEN VERSIONS IN THEIR PLACES.
It cleans up after itself, though, moving the newer dependencies back once it has finished.}
}
\description{
Run an analysis previously frozen by \code{freezr::freeze}. NOT GUARANTEED TO WORK but it does its best.

Run an analysis previously frozen by \code{freezr::freeze}. NOT GUARANTEED TO WORK but it does its best.
}
