% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tryStack}
\alias{tryStack}
\title{A way to run code, handle errors, and (unlike tryCatch) preserve the traceback.}
\usage{
tryStack(expr, silent = FALSE)
}
\description{
A way to run code, handle errors, and (unlike tryCatch) preserve the traceback.
}
\seealso{
Many thanks to Berry Boessenkool for sharing this very useful bit of code.
\url{https://stackoverflow.com/a/40899766/3371472}
}
