% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr.R
\name{freeze}
\alias{freeze}
\title{Run an analysis while freezing a copy of the code (and perhaps dependencies) for later perusal.}
\usage{
freeze(analyses_to_run, destination, run_from_cryo_storage = FALSE,
  dependencies = NULL, seed_to_set = NULL, timestamp_as_folder = TRUE,
  force = FALSE, copy_deps_kb_limit = 100, purl_aggressively = TRUE,
  chastise = TRUE, notes_file = "notes.md", other_repos_to_track = NULL)
}
\arguments{
\item{analyses_to_run}{R or R markdown files to be run and frozen.}

\item{destination}{Where to save the code (and log, and maybe dependencies and output)}

\item{run_from_cryo_storage}{If \code{FALSE} (default), runs the code from the current working directory.
If \code{TRUE}, runs from \code{destination} (or a folder within it -- see param \code{timestamp_as_folder}).
The default option, \code{FALSE}, makes it easier to write your code because you know where it'll be run from.
You can access \code{Sys.getenv()[["FREEZR_DESTINATION"]]} within your scripts to produce custom output.
The non-default option makes it easier to
break code unless it uses absolute paths throughout.}

\item{dependencies}{A list of file paths to extra files your analysis uses.
They will get saved to destination if they are smaller than \code{copy_deps_kb_limit.}}

\item{seed_to_set}{seed for rng. If \code{NULL} (default), seed doesn't get set.}

\item{timestamp_as_folder}{If \code{TRUE} (default), one extra folder is added to \code{destination}
based on the current time and date (down to the second). If \code{FALSE}, code gets saved to \code{destination}.}

\item{force}{If \code{FALSE}, refuses to continue when \code{destination} is not empty.
If \code{TRUE}, continues saving results even to non-empty \code{destination}. Default is \code{FALSE}.}

\item{copy_deps_kb_limit}{\code{dependencies} get saved to \code{destination} if they are
smaller than \code{copy_deps_kb_limit}.}

\item{purl_aggressively}{If \code{TRUE} (default), then when purling Rmd files, the
corresponding R files may be overwritten.}

\item{chastise}{If \code{TRUE} (default), creates a \code{notes.txt} file for you and nags
you about filling it.}

\item{notes_file}{Name of file to be created while \code{chastise}ing.}

\item{other_repos_to_track}{If you want to record what commit some outside repo is on -- for 
example, a package that you are developing separate from your data analysis repo -- put
the path here. You can put more than one.}
}
\value{
returns \code{destination} arg.
}
\description{
Run an analysis while freezing a copy of the code (and perhaps dependencies) for later perusal.
}
\examples{

#setwd( file.path( "~", "my_freezr_example_project", "scripts" ), )
#freeze( analyses_to_run = c( "my_functions.Rmd", "my_script.R" ),
#        destination = file.path("~", "my_freezr_example_project", "results") )
}
