% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_find}
\alias{inventory_find}
\title{Locate your inventory.}
\usage{
inventory_find(inv_location = NULL, return_existence_logical = FALSE)
}
\arguments{
\item{inv_location}{Path to the inventory you want to create, access, or modify, or to its parent folder, or to
any folder below that (as long as none of them contain other inventories). When possible, this arg
defaults to the parent of the last destination given to `freeze`.}

\item{return_existence_logical}{If TRUE, return a bool indicating whether the inventory can be found.}
}
\description{
Locate your inventory.
}
\details{
\code{inventory_*} functions help organize data as it passes through multiple stages of analysis.
 The central data structure is a table with the filename \code{.inventory.txt}. It has five
 columns: \code{tag}, \code{parent_tag}, \code{date_modified}, \code{filename}, and \code{extra}.

\code{inventory_find} helps locate the inventory.
}
