% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr_functions.R
\name{inventory_get}
\alias{inventory_get}
\title{Retrieve (paths to) important items from previous analyses.}
\usage{
inventory_get(inv_location = NULL, tag = NULL, return_all_fields = FALSE)
}
\arguments{
\item{inv_location}{Path to the inventory you want to access. If possible, this arg
defaults to the parent of the last destination given to `freeze`.}

\item{tag}{identifier for an inventory record that you want to access.}

\item{return_all_fields}{When retrieving data, return the whole inventory record instead of just the \code{filename}.}
}
\description{
Retrieve (paths to) important items from previous analyses.
}
\details{
\code{inventory_*} functions help organize data as it passes through multiple stages of analysis.
 The central data structure is a table with the filename \code{.inventory.txt}. It has five
 columns: \code{tag}, \code{parent_tag}, \code{date_modified}, \code{filename}, and \code{extra}.

\code{inventory_get} looks for this table at the \code{inv_location} you specify. Any record matching
 the tag you give will be returned.
}
