% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_save_and_add}
\alias{inventory_save_and_add}
\title{Wrapper for inventory_add that also saves tables and R objects.}
\usage{
inventory_save_and_add(
  object,
  tag = deparse(substitute(object)),
  file_type = if (is.data.frame(object)) {     "csv" } else {     "Rdata" },
  save_method = if (is.data.frame(object)) {     write.csv } else {     saveRDS },
  filename = file.path(results_path, paste0(tag, ".", file_type)),
  extra,
  inv_location = NULL,
  verbose = T,
  ...
)
}
\description{
Wrapper for inventory_add that also saves tables and R objects.
}
