% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr_functions.R
\name{run_r_or_rmd}
\alias{run_r_or_rmd}
\title{\code{source} code from an R or Rmarkdown file.}
\usage{
run_r_or_rmd(file_name, destination)
}
\arguments{
\item{file_name}{An RMarkdown file or an R file to be (\code{purl}ed and) run.
The file gets run **from the directory it is in**, not from \code{getwd()}.}
}
\value{
Name of input possibly with `.Rmd` changed to `.R`.
}
\description{
\code{source} code from an R or Rmarkdown file.
}
