% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezr.R
\name{thaw}
\alias{thaw}
\title{Run an analysis previously frozen by \code{freezr::freeze}.}
\usage{
thaw(freeze_path, thaw_path = NULL, verbose = T)
}
\arguments{
\item{freeze_path}{A directory created by \code{freezr::freeze} containing these subfolders:
- logs
- code
- output 
- user 
- dependencies (optional)}
}
\description{
Run an analysis previously frozen by \code{freezr::freeze}.
}
